// src/types.ts
export interface OWMServiceOptions {
  appid: string
  v: string
  lang: Lang
  mode: Mode
  units: Unit
}

export interface WithAppId {
  appid: string
}

export interface WithV {
  v: string
}

export interface QueryByCityName {
  q: string
}

export interface QueryByCityId {
  id: number
}

export interface QueryByGeoCoordinates {
  lat: Latitude
  lon: Longitude
}

export interface QueryByZipCode {
  zip: string
}

export type QueryParams = QueryByCityId &
  QueryByCityName &
  QueryByGeoCoordinates &
  QueryByZipCode &
  WithAppId &
  WithLangModeUnits &
  Record<string, unknown>

export interface WithLang {
  lang: Lang
}

export interface WithMode {
  mode: Mode
}

export interface WithUnits {
  units: Unit
}

export type WithLangModeUnits = WithLang & WithMode & WithUnits

export interface ByCityName {
  cityName: string
  stateCode?: string
  countryCode?: string
}

export interface ByCityId {
  cityId: number
}

export interface ByGeoGraphicCoordinates {
  lat: Latitude
  lon: Longitude
}

export interface ByZipCode {
  zipCode: string
  countryCode?: string
}

export type Unit = 'standard' | 'metric' | 'imperial'
export type Mode = 'json' | 'html' | 'xml'

export type Kelvin = number
export type Celsius = number
export type Fahrenheit = number

export type Longitude = number
export type Latitude = number

export type Lang =
  | 'af'
  | 'al'
  | 'ar'
  | 'az'
  | 'bg'
  | 'ca'
  | 'cz'
  | 'da'
  | 'de'
  | 'el'
  | 'en'
  | 'eu'
  | 'fa'
  | 'fi'
  | 'fr'
  | 'gl'
  | 'he'
  | 'hi'
  | 'hr'
  | 'hu'
  | 'id'
  | 'it'
  | 'ja'
  | 'kr'
  | 'la'
  | 'lt'
  | 'mk'
  | 'no'
  | 'nl'
  | 'pl'
  | 'pt'
  | 'pt_br'
  | 'ro'
  | 'ru'
  | 'sv, se'
  | 'sk'
  | 'sl'
  | 'sp, es'
  | 'sr'
  | 'th'
  | 'tr'
  | 'ua, uk'
  | 'vi'
  | 'zh_cn'
  | 'zh_tw'
  | 'zu'